Allow plugins to use settings utilities
There are various settings utilities included in the `omeroweb.settings`
module that would be useful to use in plugins.  Unfortunately, because
of import order and plugin settings processing that's made impossible if
they are in the same package.  The `LeaveUnset` functionality is of
particular interest.  This moves them into `omeroweb.utils` so they can
be if desired.
